# app/docs/swagger.yaml
definitions:
...
  mvideo.DeleteHistoryParam:
    properties:
      composeIds:
        items:
          type: string
        type: array
    required:
    - composeIds
    type: object
  mvideo.DeletePublishParam:
    properties:
      composeIds:
        type: string
    required:
    - composeIds
    type: object
  mvideo.VideoDetailInfo:
    properties:
      avatar:
        type: string
      barrage_num:
        type: integer
      browse_num:
        type: integer
      comment_num:
        type: integer
      cover:
        type: string
      create_at:
        type: integer
      describe:
        type: string
      fabulous_num:
        type: integer
      fans_num:
        type: integer
      is_attention:
        type: integer
      is_collect:
        type: integer
      is_like:
        type: integer
      is_recommend:
        type: integer
      is_top:
        type: integer
      labels:
        type: string
      nick_name:
        type: string
      share_num:
        type: integer
      status:
        type: integer
      title:
        type: string
      user_id:
        type: string
      video_addr:
        type: string
      video_duration:
        type: integer
      video_height:
        type: integer
      video_id:
        type: integer
      video_width:
        type: integer
    type: object
...
  mvideo.VideosInfoResp:
    properties:
      avatar:
        example: 头像
        type: string
      cover:
        example: 视频封面
        type: string
      create_at:
        example: 1600000000
        type: integer
      describe:
        example: 视频描述
        type: string
      is_attention:
        example: 1
        type: integer
      is_recommend:
        example: 1
        type: integer
      is_top:
        example: 1
        type: integer
      nick_name:
        example: 昵称
        type: string
      op_time:
        example: 1600000000
        type: integer
      title:
        example: 视频标题
        type: string
      user_id:
        example: 用户id
        type: string
      video_addr:
        example: 视频地址
        type: string
      video_duration:
        example: 1000000
        type: integer
      video_height:
        example: 1000
        type: integer
      video_id:
        example: 1000000000
        type: integer
      video_width:
        example: 1000
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    ### 一、公共参数说明（此栏参数均为Headers请求头传递）
    | 参数名 | 说明 | 示例 |
    | ------ | :----- | :----- |
    | AppId | AppId(区分Android,iOS,web) | 5EesXF1i |
    | Secret | 服务端下发的secret 通过/api/v1/client/init接口获取 调用该接口时无需传 且 不参与签名 | DnaukFwVILpcewX6 |
    | Timestamp | 请求时间戳 单位：秒 | 1588888888 |
    | Sign | 签名 | 签名 md5签名32位值 |
    | Version | 当前版本 | 1.0.1 |
    ### 二、请求体说明（此栏参数均为POST JSON传递，不可用form-data提交）

    {
    'mobileNum': '13177656222',
    'platform': 0
    }
    ### 三、接口签名生成方式
    签名加密示例:
    params = 请求的url路径(不包含域名与参数) + & + Header头参数以`&`拼接（无需按照字典序，具体看以下栗子） + & + appKey
    appKey由服务端下发 并进行保存
    sign = md5(params) 取md5 32位小写
    如：md5(/api/v1/user/mobile/login&AppId=5EesXF1i&Timestamp=1588888888&Version=1.0.1&Secret=DnaukFwVILpcewX6&RfhHecN9zsNcy19Y)
    appKey为RfhHecN9zsNcy19Y
    ### 四、API错误码文档
    [点击查看](/api/v1/doc)
    ### 五、HTTP状态码说明
    | 状态码 | 说明 |
    | ------ | :----- |
    | 200 | 操作成功 |
    | 400 | 参数错误 |
    | 500 | 内部错误 |
  license: {}
  title: 电竞社区平台（应用服）
  version: "1.0"
paths:
...
  /api/v1/video/attention:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideoDetailInfo'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 首页关注的用户发布的视频列表[分页获取] (ok)
      tags:
      - 视频模块
  /api/v1/video/browse/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideosInfoResp'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 用户浏览过的视频记录[分页获取] (ok)
      tags:
      - 视频模块
  /api/v1/video/delete/history:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 删除浏览历史记录 请求参数
        in: body
        name: DeleteHistoryParam
        required: true
        schema:
          $ref: '#/definitions/mvideo.DeleteHistoryParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 删除浏览的历史记录 (ok)
      tags:
      - 视频模块
  /api/v1/video/delete/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 删除发布的视频 请求参数
        in: body
        name: DeletePublishParam
        required: true
        schema:
          $ref: '#/definitions/mvideo.DeletePublishParam'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 删除发布的视频 (ok)
      tags:
      - 视频模块
  /api/v1/video/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 视频id
        in: query
        name: video_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mvideo.VideoDetailInfo'
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 视频详情 (ok)
      tags:
      - 视频模块
  /api/v1/video/detail/recommend:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 视频id
        in: query
        name: video_id
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideoDetailInfo'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 视频详情 (ok)
      tags:
      - 视频模块
  /api/v1/video/homepage/banner:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Banner'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 视频首页推荐的banner (ok)
      tags:
      - 视频模块
  /api/v1/video/label/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mlabel.VideoLabel'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 视频标签 (ok)
      tags:
      - 视频模块
  /api/v1/video/other/publish:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      - description: status 状态 -1 查询全部 0 审核中 1 已发布 2 不通过
        in: query
        name: status
        required: true
        type: string
      - description: 条件 -1 默认时间排序 0 播放数 1 弹幕数 2 点赞数 3 评论数 4 分享数
        in: query
        name: condition
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideosInfo'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 查看其他用户发布的视频记录[分页获取] (ok)
      tags:
      - 视频模块
  /api/v1/video/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 发布视频请求参数
        in: body
        name: VideoPublishParams
        required: true
        schema:
          $ref: '#/definitions/mvideo.VideoPublishParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"success","tm":"1588888888"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 视频发布 (ok)
      tags:
      - 视频模块
  /api/v1/video/publish/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 查看的用户id
        in: query
        name: user_id
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      - description: status 状态 -1 查询全部 0 审核中 1 已发布 2 不通过
        in: query
        name: status
        required: true
        type: string
      - description: 条件 -1 默认时间排序 0 播放数 1 弹幕数 2 点赞数 3 评论数 4 分享数
        in: query
        name: condition
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideosInfo'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 用户发布的视频记录[分页获取] (ok)
      tags:
      - 视频模块
  /api/v1/video/recommend:
    get:
      consumes:
      - application/json
      parameters:
      - description: AppId
        in: header
        name: AppId
        required: true
        type: string
      - description: 调用/api/v1/client/init接口 服务端下发的secret
        in: header
        name: Secret
        required: true
        type: string
      - description: 请求时间戳 单位：秒
        in: header
        name: Timestamp
        required: true
        type: string
      - description: 签名 md5签名32位值
        in: header
        name: Sign
        required: true
        type: string
      - default: 1.0.0
        description: 版本
        in: header
        name: Version
        required: true
        type: string
      - description: 页码 从1开始
        in: query
        name: page
        required: true
        type: string
      - description: 每页展示多少 最多50条
        in: query
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mvideo.VideoDetailInfo'
            type: array
        "500":
          description: '{"code":500,"data":{},"msg":"fail","tm":"1588888888"}'
          schema:
            type: string
      summary: 首页推荐的视频列表[分页获取] (ok)
      tags:
      - 视频模块
swagger: "2.0"